fix(sodium): sweep the shadow cull again, as the reference does
The box of 314a49d0 takes the DEFAULT state as well as ADVANCED, and at
stock settings every pack of the corpus is DEFAULT: each writes its
shadow.culling under a coloured-light conditional that is nought on every
shipped profile, so the line is never live. It also stops bounding
anything once the shadow distance reaches the render distance, which is
where the slider sits at its maximum.

Four readings on the frozen bench, one jar with the switch the only
difference, under Complementary Unbound at its stock profile: the
silhouette keeps 322 sections against the box's 875, and the frame runs
164.9 images a second against 144.4. At a denser station the same pair
was 178 against 128, so the box costs twelve to twenty-eight percent
depending on what stands around, and never nothing.

The popping the box was hiding comes back with this, and it is a defect
of ours with a cause of its own: the light's walk reuses the visibility
lists the camera's walk filled, so anything narrowing the camera's
visible set takes casters out of the map. The answer is a separate
traversal for the light, not a wider walk. vitrail/box-shadow-cull keeps
the old road reachable for a comparison.

The box never reached a release, so the changelog carries the switch and
the reading that argues for it, not a round trip no player made.
