+++ CHANGELOG.md
@@ -38,6 +38,19 @@ what the next one holds.
   What the render scale reaches and what it leaves alone is now written down, in
   `docs/render-scale.md`.
 
+- **A switch that fills the shadow map from a box around you rather than the tight
+  silhouette.** An empty file `vitrail/box-shadow-cull` in the game folder, or
+  `-Dvitrail.boxShadowCull=true`. The silhouette stays the default, which is what other loaders
+  draw. The box was tried as the default, to stop leaf blocks popping at the sun's edge as you
+  move, and measuring it is why it did not stay: on one wooded scene it drew 875 sections where
+  the silhouette draws 322, and the frame ran twelve to twenty-eight percent slower depending on
+  how much stood around. It also stops bounding anything at all once Shadow Distance reaches
+  your render distance, which is where the slider sits at its maximum.
+
+  What the box covers up is a defect of this engine and the switch is not its fix: leaves can go
+  a frame without their shadow while you walk, if you run a mod that narrows what the game
+  considers visible. That is being fixed at its cause instead.
+
 ### Changed
 
 - **Compute shaders now hit the compiled-module cache on a pack reload.** They used to miss
@@ -133,14 +146,6 @@ what the next one holds.
   outside a block, so the world dropped to vanilla. The machine symbols are now installed
   before the pack is opened, so those uniforms enter the block and the reload draws the
   pack. The toggle still reloads; it does not keep the pack on without one.
-- **Leaf shadows stay more stable while you move.** Complementary's Low profile, and any pack
-  that asks for Advanced shadow culling, now walks a box around you instead of the tight
-  silhouette along the sun. Leaves stop popping as you walk. Shadows can reach a little
-  farther than they do under Iris Advanced, and the wider walk draws more sections, so the
-  frame can cost more. The Medium profile, which asks for a safe zone, is unchanged. An
-  empty file `vitrail/swept-shadow-cull` in the game folder, or
-  `-Dvitrail.sweptShadowCull=true`, puts the old silhouette back.
-
 
 - **The button that opens a pack's settings is no longer dead until the pack has been loaded
   once.** It was live only while a pack was already drawing, so the first pack anybody picks on
